Day 3
We added the video to the site. We used the same styles on this as the sound. We then added the captions in the text and positioned them where we wanted. We made the hover effect on the top nav and then made it a transition effect. We also made the video spin forever.